Decouple GKP environment from simulator environment and clarify class structure/params

+++ simulator/operators.py
@@ -16,12 +16,10 @@ else:
 
     diag = np.diag  # k=1 option is broken in tf.linalg.diag in TF 2.1 (#35761)
 
-from simulator.utils import matrix_flatten
-
 
 def identity(N, dtype=tf.complex64):
     """Returns an identity operator in the Fock basis.
-    
+
     Args:
         N (int): Dimension of Hilbert space
         dtype (tf.dtypes.DType, optional): Returned dtype. Defaults to c64.
@@ -34,7 +32,7 @@ def identity(N, dtype=tf.complex64):
 
 def destroy(N, dtype=tf.complex64):
     """Returns a destruction (lowering) operator in the Fock basis.
-    
+
     Args:
         N (int): Dimension of Hilbert space
         dtype (tf.dtypes.DType, optional): Returned dtype. Defaults to c64.
@@ -48,15 +46,16 @@ def destroy(N, dtype=tf.complex64):
 
 def create(N, dtype=tf.complex64):
     """Returns a creation (raising) operator in the Fock basis.
-    
+
     Args:
         N (int): Dimension of Hilbert space
         dtype (tf.dtypes.DType, optional): Returned dtype. Defaults to c64.
-   
+
     Returns:
         Tensor([N, N], dtype): NxN creation operator
     """
-    return tf.cast(tf.linalg.adjoint(destroy(N, dtype=tf.complex64)), dtype=dtype)
+    # Preserve max precision in intermediate calculations until final cast
+    return tf.cast(tf.linalg.adjoint(destroy(N, dtype=tf.complex128)), dtype=dtype)
 
 
 def num(N, dtype=tf.complex64):
@@ -82,9 +81,10 @@ def position(N, dtype=tf.complex64):
     Returns:
         Tensor([N, N], dtype): NxN position operator
     """
-    sqrt2 = tf.sqrt(tf.constant(2, dtype=tf.complex64))
-    a_dag = create(N, dtype=tf.complex64)
-    a = destroy(N, dtype=tf.complex64)
+    # Preserve max precision in intermediate calculations until final cast
+    sqrt2 = tf.sqrt(tf.constant(2, dtype=tf.complex128))
+    a_dag = create(N, dtype=tf.complex128)
+    a = destroy(N, dtype=tf.complex128)
     return tf.cast((a_dag + a) / sqrt2, dtype=dtype)
 
 
@@ -93,11 +93,12 @@ def momentum(N, dtype=tf.complex64):
 
     Args:
         N (int): Dimension of Hilbert space
-    
+
     Returns:
         Tensor([N, N], c64): NxN momentum operator
     """
-    sqrt2 = tf.sqrt(tf.constant(2, dtype=tf.complex64))
-    a_dag = create(N, dtype=tf.complex64)
-    a = destroy(N, dtype=tf.complex64)
+    # Preserve max precision in intermediate calculations until final cast
+    sqrt2 = tf.sqrt(tf.constant(2, dtype=tf.complex128))
+    a_dag = create(N, dtype=tf.complex128)
+    a = destroy(N, dtype=tf.complex128)
     return tf.cast(1j * (a_dag - a) / sqrt2, dtype=dtype)
